perf(catalog): render one tab body, not both
Switching Webinars <-> Classes was visibly slow. UrlTabs writes the active tab
with router.replace, and Next 15's App Router re-renders the tree via
useSearchParams reactivity even with { scroll: false } — a lesson this repo
already learned and documented at app/explore/hooks/useUrlSyncedFilters.ts:34,
where the explore filters abandoned router.replace for window.history.
replaceState. UrlTabs never got that migration.

That router-level cause is filed separately (it is shared with Members,
Settings and ScopeToggle, and its ?page= reset semantics depend on the
re-render it would remove, so it needs its own change). This removes the
amplifier that made the catalog the worst-hit consumer.

CatalogClient passed inline renderList(...) calls as tab content, so every
re-render rebuilt BOTH kinds' full row sets, columns array and cell closures —
unpaginated, from one combined fetch. MembersTabs and SettingsTabs pass
component references instead, so the same re-render only reconciles a cheap
element descriptor and Radix mounts one panel. Catalog now does the same:
CatalogPanel owns the table, columns are memoized, and the per-kind remove
callbacks are stable so that memo is not invalidated on every parent render.

Shared row/response types move to ./types so the panel and the client agree
without one importing the other's internals.
