+++ app/dashboard/organization/[orgId]/catalog/CatalogClient.tsx
@@ -2,7 +2,7 @@
 
 import { useCallback, useMemo, useState } from "react";
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
-import { Library, Plus, Trash2, Loader2 } from "lucide-react";
+import { Library, Plus } from "lucide-react";
 
 import {
   DashboardHeader,
@@ -12,10 +12,6 @@ import { EmptyState } from "@/components/dashboard/DataCard";
 import { UrlTabs } from "@/components/dashboard/UrlTabs";
 import { Button } from "@/components/ui/button";
 import {
-  ResponsiveTable,
-  type ResponsiveColumn,
-} from "@/components/ui/responsive-table";
-import {
   Select,
   SelectContent,
   SelectItem,
@@ -23,38 +19,16 @@ import {
   SelectValue,
 } from "@/components/ui/select";
 import { useToast } from "@/hooks/use-toast";
-import { formatCurrencyAmount } from "@/utils/formatting";
 import { EventPlannerForWebinar } from "@/components/planner/components/EventPlannerForWebinar";
 import { EventPlannerForClass } from "@/components/planner/components/EventPlannerForClass";
 import type { WebinarEvent, ClassEvent } from "@/types/planner-events";
-
-type Kind = "WEBINAR" | "CLASS";
+import { CatalogPanel } from "./CatalogPanel";
+import type { CatalogResponse, Kind } from "./types";
 
 interface Expert {
   consultantProfileId: string;
   name: string;
 }
-
-/** Row shape as the API returns it — `price` is a paise string (BigInt). */
-interface CatalogRow {
-  id: string;
-  title: string;
-  price: string;
-  visibility: "PUBLIC" | "ORG_ONLY" | "ORG_AND_PUBLIC";
-  maxParticipants: number;
-  consultantProfileId: string | null;
-}
-
-interface CatalogResponse {
-  webinars: CatalogRow[];
-  classes: CatalogRow[];
-}
-
-const VISIBILITY_LABEL: Record<CatalogRow["visibility"], string> = {
-  PUBLIC: "Public",
-  ORG_ONLY: "Members only",
-  ORG_AND_PUBLIC: "Public + members",
-};
 
 export function CatalogClient({
   orgId,
@@ -132,6 +106,18 @@ export function CatalogClient({
       }),
   });
 
+  // Stable per-kind callbacks so CatalogPanel's memoized columns are not
+  // invalidated on every parent render — which would undo the point of the
+  // split. `removePlan.mutate` is referentially stable across renders.
+  const removeWebinar = useCallback(
+    (planId: string) => removePlan.mutate({ kind: "WEBINAR", planId }),
+    [removePlan],
+  );
+  const removeClass = useCallback(
+    (planId: string) => removePlan.mutate({ kind: "CLASS", planId }),
+    [removePlan],
+  );
+
   // The shared planner forms hand back a full plan object; the catalog endpoint
   // wants the flat subset it owns. Mapping here rather than widening the API
   // keeps the endpoint's contract independent of the form's internals.
@@ -180,85 +166,6 @@ export function CatalogClient({
     },
     [createPlan, expertId],
   );
-
-  const columns = useCallback(
-    (kind: Kind): ResponsiveColumn<CatalogRow>[] => [
-      {
-        key: "title",
-        header: "Offering",
-        primary: true,
-        cell: (r) => <span className="font-medium">{r.title}</span>,
-      },
-      {
-        key: "price",
-        header: "Price",
-        cell: (r) => formatCurrencyAmount(Number(r.price), "INR"),
-      },
-      {
-        key: "visibility",
-        header: "Visibility",
-        cell: (r) => VISIBILITY_LABEL[r.visibility],
-      },
-      {
-        key: "seats",
-        header: "Seats",
-        cell: (r) => r.maxParticipants,
-      },
-      {
-        key: "actions",
-        header: "",
-        hideOnCard: true,
-        cell: (r) => (
-          <Button
-            variant="ghost"
-            size="sm"
-            aria-label={`Remove ${r.title}`}
-            disabled={removePlan.isPending}
-            onClick={() => removePlan.mutate({ kind, planId: r.id })}
-          >
-            <Trash2 className="h-4 w-4" />
-          </Button>
-        ),
-      },
-    ],
-    [removePlan],
-  );
-
-  const renderList = (kind: Kind, rows: CatalogRow[]) => {
-    if (isLoading) {
-      return (
-        <div className="flex items-center justify-center py-12 text-muted-foreground">
-          <Loader2 className="mr-2 h-5 w-5 animate-spin" />
-          Loading the catalog…
-        </div>
-      );
-    }
-    if (error) {
-      return (
-        <EmptyState
-          icon={Library}
-          title="Couldn't load the catalog"
-          description={error instanceof Error ? error.message : undefined}
-        />
-      );
-    }
-    if (rows.length === 0) {
-      return (
-        <EmptyState
-          icon={Library}
-          title={`No ${kind === "WEBINAR" ? "webinars" : "classes"} yet`}
-          description="Offerings you publish here are owned by the organization and delivered by one of its experts."
-        />
-      );
-    }
-    return (
-      <ResponsiveTable
-        columns={columns(kind)}
-        rows={rows}
-        getRowId={(r) => r.id}
-      />
-    );
-  };
 
   // No experts, nothing to publish — an org plan needs somebody to deliver it,
   // so say that plainly rather than opening a form that will 422.
@@ -325,12 +232,30 @@ export function CatalogClient({
                 {
                   value: "webinars",
                   label: "Webinars",
-                  content: renderList("WEBINAR", data?.webinars ?? []),
+                  content: (
+                    <CatalogPanel
+                      kind="WEBINAR"
+                      rows={data?.webinars ?? []}
+                      isLoading={isLoading}
+                      error={error}
+                      onRemove={removeWebinar}
+                      isRemoving={removePlan.isPending}
+                    />
+                  ),
                 },
                 {
                   value: "classes",
                   label: "Classes",
-                  content: renderList("CLASS", data?.classes ?? []),
+                  content: (
+                    <CatalogPanel
+                      kind="CLASS"
+                      rows={data?.classes ?? []}
+                      isLoading={isLoading}
+                      error={error}
+                      onRemove={removeClass}
+                      isRemoving={removePlan.isPending}
+                    />
+                  ),
                 },
               ]}
             />
